#131: separate stm into its own component `StmDashboard`
